new commit

+++ src/components/pages/single/Single.js
@@ -3,11 +3,12 @@ import Sidebar from '../../sidebar/Sidebar';
 import SinglePost from '../../singlePost/SinglePost';
 import "./single.css"
 
-export default function () {
+export default function Single() {
   return (
     <div className='single'>
-        <SinglePost/>
+        <SinglePost />
         <Sidebar/>
     </div>
   )
 }
+
